Move directory manipulation out of scriptutils/__init__.py to subpackage Update packages for sbol-utilities 1.05a
